Error handling for external services-2
