+++ src/main/kotlin/uk/gov/justice/digital/assessments/restclient/AssessmentUpdateRestClient.kt
@@ -7,21 +7,15 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Component
-import org.springframework.web.reactive.function.client.ClientResponse
-import reactor.core.publisher.Mono
 import uk.gov.justice.digital.assessments.jpa.entities.AssessmentType
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.CompleteAssessmentDto
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.CreateAssessmentDto
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.CreateAssessmentResponse
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.CreateOffenderDto
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.CreateOffenderResponseDto
-import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.OASysErrorResponse
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.OasysAnswer
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.UpdateAssessmentAnswersDto
 import uk.gov.justice.digital.assessments.restclient.assessmentupdateapi.UpdateAssessmentAnswersResponseDto
-import uk.gov.justice.digital.assessments.services.exceptions.ApiClientUnknownException
-import uk.gov.justice.digital.assessments.services.exceptions.DuplicateOffenderRecordException
-import uk.gov.justice.digital.assessments.services.exceptions.UserNotAuthorisedException
 
 @Component
 class AssessmentUpdateRestClient {
@@ -46,7 +40,7 @@ class AssessmentUpdateRestClient {
       .retrieve()
       .onStatus(HttpStatus::is4xxClientError) { handleOffenderError(crn, user, it, HttpMethod.POST, path) }
       .onStatus(HttpStatus::is5xxServerError) {
-        throw ApiClientUnknownException(
+        handle5xxError(
           "Failed to create offender $crn in OASYs",
           HttpMethod.POST, path,
           ExternalService.ASSESSMENTS_UPDATE
@@ -80,7 +74,7 @@ class AssessmentUpdateRestClient {
         )
       }
       .onStatus(HttpStatus::is5xxServerError) {
-        throw ApiClientUnknownException(
+        handle5xxError(
           "Failed to create assessment for offender $offenderPK in OASYs",
           HttpMethod.POST,
           path,
@@ -117,7 +111,7 @@ class AssessmentUpdateRestClient {
         )
       }
       .onStatus(HttpStatus::is5xxServerError) {
-        throw ApiClientUnknownException(
+        handle5xxError(
           "Failed to update assessment for offender $offenderPK in OASYs",
           HttpMethod.PUT,
           path,
@@ -150,7 +144,7 @@ class AssessmentUpdateRestClient {
         handleAssessmentError(offenderPK, user, assessmentType, it, HttpMethod.PUT, path)
       }
       .onStatus(HttpStatus::is5xxServerError) {
-        throw ApiClientUnknownException(
+        handle5xxError(
           "Failed to complete assessment for offender $offenderPK in OASYs",
           HttpMethod.PUT,
           path,
@@ -159,67 +153,5 @@ class AssessmentUpdateRestClient {
       }
       .bodyToMono(UpdateAssessmentAnswersResponseDto::class.java)
       .block()
-  }
-
-  fun handleOffenderError(
-    crn: String?,
-    user: String?,
-    clientResponse: ClientResponse,
-    method: HttpMethod,
-    url: String
-  ): Mono<out Throwable?>? {
-    return when {
-      HttpStatus.CONFLICT == clientResponse.statusCode() -> {
-        log.error("Unable to create OASys offender. Duplicate OASys offender found for crn: $crn")
-        clientResponse.bodyToMono(OASysErrorResponse::class.java)
-          .map { error -> DuplicateOffenderRecordException(error.developerMessage) }
-      }
-      HttpStatus.FORBIDDEN == clientResponse.statusCode() -> {
-        log.error("Unable to create OASys offender. User $user does not have permission to create offender with crn $crn")
-        clientResponse.bodyToMono(OASysErrorResponse::class.java)
-          .map { error -> UserNotAuthorisedException(error.developerMessage) }
-      }
-      else -> handleError(clientResponse, method, url)
-    }
-  }
-
-  fun handleAssessmentError(
-    offenderPK: Long?,
-    user: String?,
-    assessmentType: AssessmentType,
-    clientResponse: ClientResponse,
-    method: HttpMethod,
-    url: String
-  ): Mono<out Throwable?>? {
-    return when {
-      HttpStatus.CONFLICT == clientResponse.statusCode() -> {
-        log.error("Unable to create OASys assessment. Existing assessment found for offender $offenderPK")
-        clientResponse.bodyToMono(OASysErrorResponse::class.java)
-          .map { error -> DuplicateOffenderRecordException(error.developerMessage) }
-      }
-      HttpStatus.FORBIDDEN == clientResponse.statusCode() -> {
-        log.error("Unable to create OASys assessment. User $user does not have permission to create assessment type: $assessmentType for offender with pk $offenderPK")
-        clientResponse.bodyToMono(OASysErrorResponse::class.java)
-          .map { error -> UserNotAuthorisedException(error.developerMessage) }
-      }
-      else -> handleError(clientResponse, method, url)
-    }
-  }
-
-  private fun handleError(clientResponse: ClientResponse, method: HttpMethod, url: String): Mono<out Throwable?>? {
-    val httpStatus = clientResponse.statusCode()
-    log.error("Unexpected exception with status $httpStatus")
-    return clientResponse.bodyToMono(String::class.java).map { error ->
-      ApiClientUnknownException(error, method, url, ExternalService.ASSESSMENTS_UPDATE)
-    }.or(
-      Mono.error(
-        ApiClientUnknownException(
-          "Unexpected exception with no body and status $httpStatus",
-          method,
-          url,
-          ExternalService.ASSESSMENTS_UPDATE
-        )
-      )
-    )
   }
 }
